move code for constructing data tables to a file
