+++ src/webgpu/capability_info.ts
@@ -3,48 +3,11 @@
 
 /* eslint-disable no-sparse-arrays */
 
-import { assertTypeTrue, ResolveType, TypeEqual, ZipKeysWithValues } from '../common/util/types.js';
+import { keysOf, makeTable, numericKeysOf, valueof } from '../common/util/data_tables.js';
+import { assertTypeTrue, TypeEqual } from '../common/util/types.js';
 import { assert, unreachable } from '../common/util/util.js';
 
 import { GPUConst } from './constants.js';
-
-type valueof<K> = K[keyof K];
-
-function keysOf<T extends string>(obj: { [k in T]: unknown }): readonly T[] {
-  return (Object.keys(obj) as unknown[]) as T[];
-}
-
-function numericKeysOf<T>(obj: object): readonly T[] {
-  return (Object.keys(obj).map(n => Number(n)) as unknown[]) as T[];
-}
-
-/**
- * Creates an info lookup object from a more nicely-formatted table. See below for examples.
- *
- * Note: Using `as const` on the arguments to this function is necessary to infer the correct type.
- */
-function makeTable<
-  Members extends readonly string[],
-  Defaults extends readonly unknown[],
-  Table extends { readonly [k: string]: readonly unknown[] }
->(
-  members: Members,
-  defaults: Defaults,
-  table: Table
-): {
-  readonly [k in keyof Table]: ResolveType<ZipKeysWithValues<Members, Table[k], Defaults>>;
-} {
-  const result: { [k: string]: { [m: string]: unknown } } = {};
-  for (const [k, v] of Object.entries<readonly unknown[]>(table)) {
-    const item: { [m: string]: unknown } = {};
-    for (let i = 0; i < members.length; ++i) {
-      item[members[i]] = v[i] ?? defaults[i];
-    }
-    result[k] = item;
-  }
-  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
-  return result as any;
-}
 
 // Queries
